example 13-cnn-inference: kernels.rs + reference impl + reference.bin (TASK-0053 cycle-2)
Lands the cycle-2 deliverable now that TASK-0209 has cleared the
partial-sub-array codegen blocker.

Algorithm changes (per docs/reference-impl-policy.md §3, all three move
together):
- prog.algo.nuc: switch element type from f32 to i32. The CNN
  demonstrates layer-wise dataflow shape (PRD §9 row 13), not real
  quantised neural-net inference; integer arithmetic is bit-deterministic
  by Rust's language definition (PRD §13 "Leaning toward integer-only
  for v2"). f32 would have required spelling out a reduction order in
  BOTH kernels.rs and reference/, with cross-architecture risk from FMA
  contraction; i32 avoids all of that for the same demonstration value.
- kernels.rs: implement all six kernels (conv_block_1 / conv_block_2 /
  classifier as pure; load_input / save_output as effectful). 3x3 SAME-
  padded conv with zero-pad; ReLU; 2x2 maxpool stride 2; dense classifier
  with per-class wrapping dot product. Weights are deterministic integer
  formulae keyed by (oc, ic, ky, kx) / (class, k). Range analysis pinned
  in the docstring; all accumulators well inside i32::MAX with the
  committed input pattern.
- reference/Cargo.toml + src/main.rs: independent std-only Rust crate
  per policy §2. Same algorithm, separate source: shape constants
  re-declared, weight formulae re-typed, control structure expressed
  via a generic `forward_conv_pool` (kernels.rs uses two separate
  conv1/conv2 bodies). A bug in one file is not automatically
  duplicated in the other.
- input.bin: 50176 bytes = B*C0*H*W = 12544 LE i32 words. Generated by
  a deterministic per-(n,c,y,x) integer formula documented in README
  and inlined verbatim.
- reference.bin: 640 bytes = B*N_CLASSES = 160 LE i32 logits.
  Regenerated by running the reference impl on input.bin.
- README.md: new sections covering required-schedule status table
  (naive PASS on both backends, batch_parallel/pipeline_parallel
  deferred with task pointers), determinism strategy (i32 with explicit
  rationale tied to PRD §13), I/O format with regeneration commands,
  honest limitations (no training, no quantisation, no bias, no softmax,
  hand-crafted not learned weights).

Classifier weight modulus design note: an earlier `% 5` modulus folded
classes 0..4 onto 5..9 in the logit output (any polynomial f(c) mod 5
has period 5; this includes the linear class*131 term since 131 % 5 ==
1). Settled on `% 11` (smallest M >= N_CLASSES=10 making class*131 mod
M injective on 0..9, since 131 is coprime with 11). Documented inline
in both kernels.rs and reference/src/main.rs.

Matrix changes (e2e-matrix.toml):
- 13-cnn-inference added to runnable_examples.
- naive × pthreads-sync and naive × mp-tcp-bufsync added as REQUIRED
  cells (milestone M3); both byte-identical to reference.bin and to
  each other (single-host => mp-tcp single-process => shared single-
  worker renderer => byte-identical arithmetic).
- batch_parallel × pthreads-sync: SKIP with TASK-0211 reason. New
  follow-up: partition=workers replicates the loop body into per-worker
  scopes, but only w0 receives `input` via slot_0.wait(); the emitted
  partial-sub-array reads in w1/w2/w3 reference an undeclared `input`
  local => E0425. Pre-existing multi-worker transfer-distribution gap
  exposed by TASK-0209's sub-array codegen; the scalar-only multi-worker
  path emitted by the same bug for examples 01-07 happens not to
  trigger E0425 because no current required cell combines
  partition=workers + multi-worker recv into one scope.
- batch_parallel × mp-tcp-bufsync: SKIP with TASK-0175 + TASK-0211.
- pipeline_parallel × both: SKIP with TASK-0210 reason
  (async + buffer=3 + notify=event capability mismatch).

Test fixups:
- compiler/tests/algo_lower.rs::lowers_example_13_cnn_inference:
  ScalarType::F32 -> ScalarType::I32 in the feat1 assertion to match
  the new prog.algo.nuc element type.
- backends/pthreads-sync/tests/emit.rs::partial_index_lowers_to_sub_slice:
  stub kernels.rs in the test now uses Vec<i32>; comments updated.
  Substring assertions on the emitted main.rs are unchanged (the
  sub-slice expression text does not vary with element type).

Verification gate (all green):
1. just test                  : 469/0/2 passed (matches pre-cycle baseline)
2. cargo clippy --all-targets : clean -D warnings
3. just e2e                   : 36 / 28 PASS / 0 FAIL / 8 SKIPPED / 0 required-fail
                                (delta from pre-cycle 30/26/0/4/0: +6 cells, +2 pass, +4 skip)
4. just determinism-check (x2): all cells byte-identical across both runs
5. just determinism-check-negative : NUC_NONDET_PERTURBED_CELLS=28; bites
6. just xbackend-check-negative    : NUC_XBACKEND_CORRUPTED_DETECTED=1; bites
7. just ci                    : exit 0

Honest limitations / not-done-this-cycle:
- batch_parallel + pipeline_parallel deferred per cycle-2 scope.
  TASK-0211 (new) tracks the multi-worker transfer-distribution gap;
  TASK-0210 (pre-existing) tracks the async/event capability gap.
- Cross-architecture i32 determinism is by Rust language definition;
  not separately exercised here (no foreign-arch CI yet).
- No training, no quantisation, no bias, no softmax — all documented in
  README under "Honest limitations".

+++ nucleus/backends/pthreads-sync/tests/emit.rs
@@ -263,10 +263,12 @@ fn kernels_rs_is_copied_verbatim() {
 /// against the real example 13 algorithm so the test exercises the
 /// real lowering path — not a synthetic Fire that might encode the
 /// shape differently from what AlgoIR produces. A scratch
-/// `kernels.rs` with the `(Vec<f32>) -> Vec<f32>` aggregate signature
-/// matches the kernel-param convention picked under TASK-0103; the
-/// contract pass reports the aggregate gap (TASK-0012) and the
-/// driver proceeds, exactly as it does on the command line.
+/// `kernels.rs` with the `(Vec<i32>) -> Vec<i32>` aggregate signature
+/// matches the kernel-param convention picked under TASK-0103 AND
+/// matches the i32 element type of example 13's `prog.algo.nuc` post
+/// TASK-0053 cycle-2; the contract pass reports the aggregate gap
+/// (TASK-0012) and the driver proceeds, exactly as it does on the
+/// command line.
 #[test]
 fn partial_index_lowers_to_sub_slice() {
     let root = repo_root();
@@ -276,6 +278,10 @@ fn partial_index_lowers_to_sub_slice() {
     // Synthesise a stub kernels.rs for example 13 (the real one is
     // TASK-0053's scope; we only need a path whose signatures are
     // contract-compatible enough for the driver path to proceed).
+    // Element type is `i32` to match the example's `prog.algo.nuc`
+    // after TASK-0053 cycle-2 landed (CNN uses integer arithmetic
+    // per PRD §13 "Leaning toward integer-only for v2"; see
+    // examples/13-cnn-inference/README.md).
     let kernels_path = scratch.join("kernels.rs");
     fs::write(
         &kernels_path,
@@ -287,11 +293,11 @@ const C0: usize = 1;
 const C1: usize = 8;
 const C2: usize = 16;
 const N_CLASSES: usize = 10;
-pub fn load_input() -> Vec<f32> { vec![0.0; B * C0 * H * W] }
-pub fn save_output(_data: Vec<f32>) {}
-pub fn conv_block_1(_x: Vec<f32>) -> Vec<f32> { vec![0.0; C1 * (H / 2) * (W / 2)] }
-pub fn conv_block_2(_x: Vec<f32>) -> Vec<f32> { vec![0.0; C2 * (H / 4) * (W / 4)] }
-pub fn classifier(_x: Vec<f32>) -> Vec<f32> { vec![0.0; N_CLASSES] }
+pub fn load_input() -> Vec<i32> { vec![0; B * C0 * H * W] }
+pub fn save_output(_data: Vec<i32>) {}
+pub fn conv_block_1(_x: Vec<i32>) -> Vec<i32> { vec![0; C1 * (H / 2) * (W / 2)] }
+pub fn conv_block_2(_x: Vec<i32>) -> Vec<i32> { vec![0; C2 * (H / 4) * (W / 4)] }
+pub fn classifier(_x: Vec<i32>) -> Vec<i32> { vec![0; N_CLASSES] }
 "#,
     )
     .expect("write stub kernels.rs");
@@ -312,18 +318,18 @@ pub fn classifier(_x: Vec<f32>) -> Vec<f32> { vec![0.0; N_CLASSES] }
     let main_rs = fs::read_to_string(&result.main_rs).unwrap();
 
     // ---- Argument-side: `input[n]` (rank-1 index on rank-4 data) ----
-    // Pre-init confirmed: `input` is rank-4 f32[16][1][28][28] = 12544
+    // Pre-init confirmed: `input` is rank-4 i32[16][1][28][28] = 12544
     // slots, sub-array per outer index = 1*28*28 = 784 slots.
     assert!(
         main_rs.contains("input[((n) * 784) as usize..((n) * 784) as usize + 784usize].to_vec()"),
         "TASK-0209 AC#1: partial-index argument did NOT lower to a sub-slice\
          `.to_vec()`; pre-TASK-0209 emission `input[(n) as usize]` (single \
-         f32) would have been emitted, breaking cargo build (E0308 expected \
-         Vec<f32>, found f32). Generated main.rs:\n{main_rs}"
+         i32) would have been emitted, breaking cargo build (E0308 expected \
+         Vec<i32>, found i32). Generated main.rs:\n{main_rs}"
     );
 
     // ---- Output-side: `feat1[n]` (rank-1 LHS on rank-4 data) ----
-    // feat1 is f32[16][8][14][14] = 25088 slots, sub-array per outer
+    // feat1 is i32[16][8][14][14] = 25088 slots, sub-array per outer
     // index = 8*14*14 = 1568. Cycle-2 review-gate finding-2: the
     // emission now binds the RHS to `let _rhs = ...;`, asserts
     // `_rhs.len() == sub_len` for fail-loud-with-context on kernel
@@ -343,7 +349,7 @@ pub fn classifier(_x: Vec<f32>) -> Vec<f32> { vec![0.0; N_CLASSES] }
          the bound-RHS + length-assert + sub-range `.copy_from_slice(&_rhs)` \
          emission. Pre-TASK-0209 cycle-1 would have emitted scalar \
          `feat1[(n) as usize] = kernels::conv_block_1(...)` breaking cargo \
-         build (E0308 expected f32, found Vec<f32>); pre-cycle-2 would have \
+         build (E0308 expected i32, found Vec<i32>); pre-cycle-2 would have \
          emitted the bare `.copy_from_slice(&kernels::conv_block_1(...))` \
          form which panics with std's terse message on kernel-impl length \
          mismatch. Generated main.rs:\n{main_rs}"
@@ -367,7 +373,7 @@ pub fn classifier(_x: Vec<f32>) -> Vec<f32> { vec![0.0; N_CLASSES] }
     // ---- AC#3: the emitted crate actually cargo-builds. The driver
     //      emit() above wrote it; verify with cargo check (not full
     //      build — the test only needs to prove the type-checker
-    //      accepts the Vec<f32>/Vec<f32> contract).
+    //      accepts the Vec<i32>/Vec<i32> contract).
     //
     // We isolate the target dir under `scratch` to avoid colliding
     // with the workspace target. CARGO_TARGET_DIR points the build
